backend/requirements.txt
# VectorShift Integrations Backend Dependencies - Updated 2025
# Core web framework and API
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# HTTP clients and networking
httpx==0.28.1
requests==2.32.4
httpcore==1.0.9
httptools==0.6.4
h11==0.16.0
websockets==15.0.1

# Data validation and serialization
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0

# Database and caching
redis==6.3.0

# Environment and configuration
python-dotenv==1.1.1
python-multipart==0.0.20

# Async and utilities
anyio==4.10.0
sniffio==1.3.1
watchfiles==1.1.0
typing-extensions==4.14.1
typing-inspection==0.4.1

# Standard libraries
click==8.2.1
colorama==0.4.6
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
PyYAML==6.0.2